refactor(aws-sdk-helpers): Rename create-pre-signed-url to signed-url
